refactor: move load to separate file
