+++ cmd/vela-cli/main.go
@@ -9,7 +9,6 @@ import (
 	"os"
 	"time"
 
-	"github.com/go-vela/cli/action/config"
 	"github.com/go-vela/cli/internal/client"
 	"github.com/go-vela/cli/version"
 
@@ -149,51 +148,4 @@ func main() {
 	if err != nil {
 		logrus.Fatal(err)
 	}
-}
-
-// load is a helper function that loads the necessary configuration for the CLI.
-func load(c *cli.Context) error {
-	// create the config file configuration
-	//
-	// https://pkg.go.dev/github.com/go-vela/cli/action/config?tab=doc#Config
-	conf := &config.Config{
-		Action: "load",
-		File:   c.String("config"),
-	}
-
-	// validate config file configuration
-	//
-	// https://pkg.go.dev/github.com/go-vela/cli/action/config?tab=doc#Config.Validate
-	err := conf.Validate()
-	if err == nil {
-		// execute the load call for the config file configuration
-		//
-		// https://pkg.go.dev/github.com/go-vela/cli/action/config?tab=doc#Config.Load
-		err = conf.Load(c)
-		if err != nil {
-			return err
-		}
-	}
-
-	// set log level for the CLI
-	switch c.String("log.level") {
-	case "t", "trace", "Trace", "TRACE":
-		logrus.SetLevel(logrus.TraceLevel)
-	case "d", "debug", "Debug", "DEBUG":
-		logrus.SetLevel(logrus.DebugLevel)
-	case "w", "warn", "Warn", "WARN":
-		logrus.SetLevel(logrus.WarnLevel)
-	case "e", "error", "Error", "ERROR":
-		logrus.SetLevel(logrus.ErrorLevel)
-	case "f", "fatal", "Fatal", "FATAL":
-		logrus.SetLevel(logrus.FatalLevel)
-	case "p", "panic", "Panic", "PANIC":
-		logrus.SetLevel(logrus.PanicLevel)
-	case "i", "info", "Info", "INFO":
-		fallthrough
-	default:
-		logrus.SetLevel(logrus.InfoLevel)
-	}
-
-	return nil
 }
